+++ packages/toprf-secure-backup/CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- feat: Add auth token error handling (expired and invalid) to TOPRF eval request in `recoverTOPRFSeed` ([#147](https://github.com/MetaMask/toprf-secure-backup/pull/147))
+
 ## [0.8.0]
 
 ### Changed